+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [0.5.0]
 ### Added
 - Signal rail that visualizes what happens in the stack, shown on the left side
   of each unit in the rack.
@@ -312,7 +312,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - a command line utility to convert .yml songs to .asm
   - a command line utility to play the songs on command line
 
-[Unreleased]: https://github.com/vsariola/sointu/compare/v0.4.1...HEAD
+[Unreleased]: https://github.com/vsariola/sointu/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/vsariola/sointu/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/vsariola/sointu/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/vsariola/sointu/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/vsariola/sointu/compare/v0.2.0...v0.3.0